final push

+++ controller/event_controller.go
@@ -1,19 +1,19 @@
 package controller
 
 import (
-	"eliferden.com/restapi/models"
 	"fmt"
-	"github.com/gin-gonic/gin"
-	"log"
 	"net/http"
 	"strconv"
+
+	"eliferden.com/restapi/models"
+	"github.com/gin-gonic/gin"
 )
 
 func GetEvents(context *gin.Context) {
 	events, err := models.GetAllEvents()
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not fetch events, try again later."})
-		log.Fatal(err)
+		 
 		return
 	}
 	context.JSON(http.StatusOK, events)
@@ -23,14 +23,14 @@ func GetEvent(context *gin.Context) {
 	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not parse event id."})
-		log.Fatal(err)
+		 
 		return
 	}
 
 	event, err := models.GetEventByID(eventId)
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message: ": "There was an error while fetching the event"})
-		log.Fatal(err)
+		 
 		return
 	}
 
@@ -38,6 +38,8 @@ func GetEvent(context *gin.Context) {
 }
 
 func CreateEvent(context *gin.Context) {
+	
+
 	var event models.Event
 	err := context.ShouldBindJSON(&event)
 	if err != nil {
@@ -45,7 +47,7 @@ func CreateEvent(context *gin.Context) {
 		fmt.Print(err)
 		return
 	}
-	event.UserID = 1
+	event.UserID = context.GetInt64("userId")
 
 	err = event.Save()
 	if err != nil {
@@ -61,16 +63,24 @@ func UpdateEvent(context *gin.Context) {
 	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not parse event id."})
-		log.Fatal(err)
 		return
 	}
 
-	_, err = models.GetEventByID(eventId)
+	userId := context.GetInt64("userId")
+	event , err := models.GetEventByID(eventId)
+
+	if event.UserID != userId {
+		context.JSON(http.StatusUnauthorized, gin.H{"message" : "Not authorized to update event"})
+		return 
+	}
+
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "There was an error while fetching the event"})
-		log.Fatal(err)
+		 
 		return
 	}
+	
+
 	var updatedEvent models.Event
 
 	err = context.ShouldBindJSON(&updatedEvent)
@@ -84,7 +94,7 @@ func UpdateEvent(context *gin.Context) {
 	err = updatedEvent.UpdateEvent()
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "There was an error while updating the event"})
-		log.Fatal(err)
+		 
 		return
 	}
 	context.JSON(http.StatusOK, gin.H{"message: ": "Event updated successfully"})
@@ -94,23 +104,79 @@ func DeleteEvent(context *gin.Context) {
 	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not parse event id."})
-		log.Fatal(err)
+		 
 		return
 	}
 
-	event, err := models.GetEventByID(eventId)
+	userId := context.GetInt64("userId")
+	event , err := models.GetEventByID(eventId)
+
+	if event.UserID != userId {
+		context.JSON(http.StatusUnauthorized, gin.H{"message" : "Not authorized to delete event"})
+		return 
+	}
+
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "There was an error while fetching the event"})
-		log.Fatal(err)
+		 
 		return
 	}
 	err = event.DeleteEvent(eventId)
 
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "There was an error while deleting the event"})
-		log.Fatal(err)
+		 
 		return
 	}
 	context.JSON(http.StatusOK, gin.H{"message: ": "Event deleted successfully"})
 
+}
+
+func RegisterForEvent (context *gin.Context) {
+	userId := context.GetInt64("userId")
+	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not parse event id."})
+		 
+		return
+	}
+
+	event, err := models.GetEventByID(eventId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "Could not find the event."})
+		return
+	}
+
+	err = event.Register(userId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not register user for event"})
+		return
+	}
+
+	context.JSON(http.StatusCreated, gin.H{"message" : "Event is registered"})
+
+}
+
+func CancelRegistration (context *gin.Context) {
+	userId := context.GetInt64("userId")
+	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message: ": "Could not parse event id."})
+		 
+		return
+	}
+
+	event, err := models.GetEventByID(eventId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message: ": "Could not find the event."})
+		return
+	}
+
+	err = event.Unregister(userId, eventId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not unregister event"})
+		return
+	}
+
+	context.JSON(http.StatusCreated, gin.H{"message" : "Event is unregistered"})
 }
